Delete or hide scrabble .js
Afternoon, please i'm beggin you if you could kindly delete or hide this on the 25th of July to the 1st of August, because I will like to use it to assist me in the program i'm making for my college work, that would be very nice of you. If you don't want to then It's perfectly understanable, its just that I would get in serious trouble for plaigirism.